+++ dataprovider/foxsports/model/mllb_pitching_box_score_stats.go
@@ -7,13 +7,13 @@ type MLBPitchingBoxScoreStats struct {
 	// PullTimestamp is the fetch timestamp for when the request was made to the API
 	PullTimestamp time.Time `json:"pull_timestamp"`
 	// PullTimestampParquet is the fetch timestamp (in milliseconds)
-	PullTimestampParquet int64 `parquet:"name=pull_timestamp, type=INT64, logicaltype=TIMESTAMP, logicaltype.unit=MILLIS, logicaltype.isadjustedtoutc=true, convertedtype=TIMESTAMP_MILLIS"`
+	PullTimestampParquet int64 `json:"-" parquet:"name=pull_timestamp, type=INT64, logicaltype=TIMESTAMP, logicaltype.unit=MILLIS, logicaltype.isadjustedtoutc=true, convertedtype=TIMESTAMP_MILLIS"`
 	// EventID is a unique ID that maps to the matchup e.g. 86833
 	EventID int64 `json:"event_id" parquet:"name=event_id, type=INT64"`
 	// EventTime is the timestamp associated with the matchup
 	EventTime time.Time `json:"event_time"`
 	// EventTimeParquet is the timestamp associated with the matchup (in milliseconds)
-	EventTimeParquet int64 `parquet:"name=event_time, type=INT64, logicaltype=TIMESTAMP, logicaltype.unit=MILLIS, logicaltype.isadjustedtoutc=true, convertedtype=TIMESTAMP_MILLIS"`
+	EventTimeParquet int64 `json:"-" parquet:"name=event_time, type=INT64, logicaltype=TIMESTAMP, logicaltype.unit=MILLIS, logicaltype.isadjustedtoutc=true, convertedtype=TIMESTAMP_MILLIS"`
 	// TeamID is the player's team's ID e.g. 21
 	TeamID int64 `json:"team_id" parquet:"name=team_id, type=INT64"`
 	// Team is the player's team name e.g. Atlanta Braves
@@ -28,7 +28,7 @@ type MLBPitchingBoxScoreStats struct {
 	Player string `json:"player" parquet:"name=player, type=BYTE_ARRAY"`
 	// Record is starting pitcher's record
 	Record *string `json:"record" parquet:"name=record, type=BYTE_ARRAY"`
-	// PitchingOrder - The sequence of pitchers who played during the event per team
+	// PitchingOrder - The sequence of pitchers who played during the event per team (starting from 1)
 	PitchingOrder int32 `json:"pitching_order" parquet:"name=pitching_order, type=INT32"`
 	// InningsPitched (IP) - https://www.mlb.com/glossary/standard-stats/innings-pitched
 	InningsPitched float32 `json:"innings_pitched" parquet:"name=innings_pitched, type=FLOAT"`